refactor: application 패키지 구성 및 VoucherApplication 이름 변경
기존 상위 모듈인 CommandType 이넘과 
VoucherApplication을 application 패키지로 이동
VoucherApplication 이름을 ConsoleApplication으로 변경
